requirements.txt
Flask
Flask-SQLAlchemy
pymupdf
openai>=1.40.0
google-generativeai
markdown
bleach
langdetect
requests
google-auth
SQLAlchemy
psycopg2-binary
python-docx
reportlab
gunicorn
